REFACTOR: moved checker definitions
